requirement.txt
autoreject==0.4.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
eeglabio==0.0.2.post4
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
h5io==0.2.2
h5py==3.11.0
idna==3.6
importlib_resources==6.1.3
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.11.2
lion-pytorch==0.1.4
MarkupSafe==2.1.5
matplotlib==3.8.3
mkl==2021.4.0
mne==1.5.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
pymatreader==0.0.32
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchmetrics==1.3.2
torchsummary==1.5.1
torchvision==0.18.0+cu121
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
xmltodict==0.13.0
zipp==3.17.0
